+++ GameAnarchy/Manager/CityServicesManager.cs
@@ -1,97 +1,7 @@
-﻿namespace GameAnarchy;
+﻿namespace GameAnarchy.Manager;
 using ColossalFramework;
-using ColossalFramework.UI;
-using ICities;
 using System;
 using UnityEngine;
-
-internal class OptionsPanelCategoriesManager {
-    public static float MainPanelWidth = 1074f;
-    public static float CategoriesDefaultWidth = 268;
-    public static float ContainerDefaultWidth = 764;
-    public static float ContainerDefaultPosX = 296;
-    public static float ContainerDefaultPosY = 54;
-    public static void SetCategoriesOffset() => SetCategoriesOffset(UIView.library.Get<UIPanel>("OptionsPanel"));
-    public static void SetCategoriesOffset(UIComponent component) {
-        var categories = component.Find<UIListBox>("Categories");
-        var optionsContainer = component.Find<UITabContainer>("OptionsContainer");
-        var delta = Config.Instance.OptionPanelCategoriesHorizontalOffset + CategoriesDefaultWidth;
-        var panelTotalWidth = MainPanelWidth + Config.Instance.OptionPanelCategoriesHorizontalOffset;
-        component.width = panelTotalWidth;
-        categories.width = delta;
-        optionsContainer.width = ContainerDefaultWidth;
-        optionsContainer.relativePosition = new Vector2(ContainerDefaultPosX + Config.Instance.OptionPanelCategoriesHorizontalOffset, ContainerDefaultPosY);
-    }
-}
-
-internal class AchievementsManager {
-    private static UIPanel unlockingPanel;
-    private static UIButton button;
-    private static bool isInGame;
-    public static void InitializeAchievements(LoadMode loadMode) {
-        isInGame = true;
-        if (loadMode == LoadMode.NewGame || loadMode == LoadMode.LoadGame || loadMode == LoadMode.NewGameFromScenario || loadMode == LoadMode.LoadScenario) {
-
-            unlockingPanel = UIView.Find<UIPanel>("UnlockingPanel");
-            if (unlockingPanel is null) {
-                ExternalLogger.Log("Initialize achievements failed, couldn't find UnlockingPanel.");
-            } else {
-                ExternalLogger.Log($"{loadMode} mode, start initalize UnlockingPanel.");
-                var tabstrip = unlockingPanel.Find<UITabstrip>("Tabstrip");
-                button = tabstrip.Find<UIButton>("Achievements");
-            }
-            UpdateAchievements(Config.Instance.EnabledAchievements);
-        } else {
-            ExternalLogger.Log("Not Game mode, do not initalize UnlockingPanel.");
-            UpdateAchievements(Config.Instance.EnabledAchievements);
-        }
-    }
-
-    public static void UpdateAchievements(bool isEnable) {
-        if (!isInGame) {
-            return;
-        }
-        try {
-            if (isEnable) {
-                Singleton<SimulationManager>.instance.m_metaData.m_disableAchievements = SimulationMetaData.MetaBool.False;
-            } else {
-                Singleton<SimulationManager>.instance.m_metaData.m_disableAchievements = SimulationMetaData.MetaBool.True;
-            }
-            if (button is not null) {
-                button.isEnabled = isEnable;
-            }
-            ExternalLogger.Log($"Update achievements succeed, status: {isEnable}");
-        } catch (Exception e) {
-            ExternalLogger.Log($"Update achievements status failure, detail: {e.Message}");
-        }
-    }
-
-    public static void Destroy() {
-        button = null;
-        unlockingPanel = null;
-        isInGame = false;
-    }
-}
-
-internal class FireControlManager {
-    public static uint buildingFireSpreadCount;
-    public static uint buildingFireSpreadAllowed;
-    public static uint treeFireSpreadCount;
-    public static uint treeFireSpreadAllowed;
-
-    public static bool GetFireProbability(uint probaility, ref uint count, ref uint allowed) {
-        if (Singleton<SimulationManager>.exists) {
-            var randomValue = Singleton<SimulationManager>.instance.m_randomizer.UInt32(1, 100);
-            count++;
-            if (randomValue < probaility) {
-                allowed++;
-            }
-            return randomValue <= probaility;
-        }
-        ExternalLogger.Error("SimulationManager does not exist.");
-        return false;
-    }
-}
 
 internal class CityServicesManager {
     public static ushort numBuildingChunks = 32;
