+++ GameAnarchy/Patches/OptionsMainPanelPatch.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Reflection.Emit;
 using System.IO;
+using GameAnarchy.Manager;
 
 public static class OptionsMainPanelPatch {
     public static MethodInfo GetOriginalOnVisibilityChanged() => AccessTools.Method(typeof(OptionsMainPanel), "OnVisibilityChanged");
