+++ GameAnarchy/Properties/Localize.resx
@@ -144,6 +144,12 @@
   <data name="BuildingSpreadFireProbability" xml:space="preserve">
     <value>Building spread fire probability</value>
   </data>
+  <data name="CityBankruptcyWarningThreshold" xml:space="preserve">
+    <value>City bankruptcy warning threshold</value>
+  </data>
+  <data name="CityBankruptcyWarningThresholdMinor" xml:space="preserve">
+    <value>Game default value is -10000</value>
+  </data>
   <data name="CityServiceOptions" xml:space="preserve">
     <value>City Service Options</value>
   </data>
@@ -355,7 +361,7 @@
     <value>Remove museum fire</value>
   </data>
   <data name="RemoveNotEnoughMoney" xml:space="preserve">
-    <value>Remove not enough money warning</value>
+    <value>Remove not enough money limitation</value>
   </data>
   <data name="RemoveNotEnoughMoneyMinor" xml:space="preserve">
     <value>When “Money Anarchy” or "Unlimited Money" is enabled, this option is enabled automatically enforced.</value>
